feature : 패키지 depth 1 추가
1. 각 패키지를 관리하는 Ladder 클래스를 추가했습니다.

end
