+++ src/main/java/Ladder/domain/Element.java
@@ -1,4 +1,4 @@
-package domain;
+package Ladder.domain;
 
 public class Element {
