+++ src/main/java/Ladder/view/InputView.java
@@ -1,4 +1,4 @@
-package view;
+package Ladder.view;
 
 import java.util.Arrays;
 import java.util.List;
